requirements.txt
discord.py[voice]
pip
libffi-devel
python3.5-dev
asyncio
requests
